Passed onTaskCreated parameter to createTask thunk
Replaced `familyCode` string parameter with `onTaskCreated`
function parameter to createTask thunk. The familyCode string
will be retrieved with Redux getState(). Inside the last then()
method the `onTaskCreated` function will be called.

The `onTaskCreated` function is passed as a prop
to the CreateTaskForm component, when it's rendered
inside the Dashboard component. The function is used
to determine if the CreateTaskForm component is displayed.