Added comments for conditions to render Prize UI
Notes about JSX curly braces inside the <div className="userPrize">

1. The First pair of braces will render UI, the string Prize name,
if a `userPrize` property exists in the store.
If the `userPrize` property exists, both operands evaluate as `truthy`
which makes the `&&` expression evaluate to True. This returns the
Prize name as a string.

2. The Second pair of braces will render the Delete prize button if
the store contains a `userPrize` property. If the prize exists, there
should be a rendered button to delete the Prize.
The `&&` expression will evaluate to True. This returns a button element
with a function to delete the Prize.