Completed cancel button for create task form
The prop `onTaskCreated` is a method passed to the CreateTaskForm
component. The method changes the `isAddTaskFormVisible`
state so that the CreateTaskForm component is not visible.

The state `isAddTaskFormVisible` changes from True to False when:

1. The Create Task form is submitted, and there's a dispatch to
the `createTask` thunk with `onTaskCreated` as an argument.
Inside the thunk, after the dispatch of Create Task Success,
the `onTaskCreated` method is called to change the state
of `isAddTaskFormVisible` to False.
* After the submission & creation of a Task, the form is hidden.

2. The prop `onTaskCreated` method is also used to hide
the Create Task form if the user decides to Cancel the creation
of a new Task. The user clicks the Cancel button, and
the `onTaskCreated` is called, changing the state of
`isAddTaskFormVisible` to False to hide the form.