State decided if button says create or update
The CreateOrUpdatePrizeForm component has access to the state
using connect() and mapStateToProps. If `userPrize` exists
in state, the form to update the prize will have a button
that displays `Cancel Update`. If the `userPrize` does not exist,
the form to create a prize will have a button that displays
`Cancel Create`.