Child user must be logged in for refresh auth token to run
The condition of a logged in Child user must be true
for the setInterval function to run. If the Child user's
`tasksReadyForReview` is true, then a dispatch to refresh
the auth token will run continuously.

Refreshing the Auth token is an "ugly" way for the Parent
to check the state of `tasksReadyForReview` for each Child user
in `state.tasks.allChildStatus`. Checking the state
of `tasksReadyForReview` allows the approve tasks button
in the ChildStatusList component to be enabled,
so the Parent can approve each Child's task list.