Poll for Childs tasksReadyForReview, so Parent button is active
- When Parent logged in:
Created `pollGetChildStatus` thunk so the ChildStatusList
Parent component can continously check for the each Child's
`tasksReadyForReview` in `state.tasks.allChildStatus`.
The property is part of the condition to enable the Parent's
approve child's tasks button, for each child.

- When Child logged in:
Dispatch `refreshAuthToken` from `notifyParentTasksReadyForReview`
thunk, so that the Child's `tasksReadyForReview` is updated
on the Child user's database and `state.auth.currentUser`.