Edited pollGetChildStatus, made stopPollGetChildStatus
Edited pollGetChildStatus thunk to dispatch SET_INTERVAL_ID
to the store. The interval Id needs to be saved in the state,
because later the setInterval called in pollGetChildStatus
will need to be cleared (stopped) when the user logs out.
This is to prevent errors and crashing.

When the user clicks the log out button, the stopPollGetChildStatus
thunk is dispatched. The interval Id is taken from state
and passed to clearInterval() to stop the setInterval called
in pollGetChildStatus.