Used redux to get errors for role validation when creating User
Check the radio button validation for role at the level of
the RegistrationForm component during the submission of
creating a new user. Removed field level validation on the
radio buttons.

The RegistrationForm is passed as a component to `reduxForm` HOC.
Have not figured how to access `state.form.registration.syncErrors`
to check the role validation using reduxForm. So pass the
reduxForm connected RegistrationForm to Redux itself via `connect`
in order to access syncErrors state. This is a hack.
There should be a way to use `reduxForm` alone, without
using `connect`.