Tested `Create Task` button click shows CreateTaskForm
Tested `Create Task` button click shows CreateTaskForm.

Used `mount` to simulate button click, to render CreateTaskForm.
Needed to mock the Redux Store methods getState, dispatch, and subscribe
to get the Enzyme test to work. Also needed to import Provider so
the component can connect to the mock store.

The getState method needs to set: `state.auth.currentUser.role`
value to `parent` in order for the ShowIfRoleIs HOC to display
the appropriate button and CreateTaskForm for ONLY the parent.